Moved pugh_torch load_state_dict method into a mixin LoadStateDictMixin
